Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/dom/dynamic-markup-insertion/opening-the-input-stream/008-expected.txt


FAIL Replacement of document prototype object after document.open assert_equals: expected (undefined) undefined but got (number) 1
